refactor: move URI method from LBWS to uri.go
