+++ internal/pkg/describe/lb_web_service_test.go
@@ -16,200 +16,10 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestLBWebServiceURI_String(t *testing.T) {
-	testCases := map[string]struct {
-		dnsNames []string
-		path     string
-		https    bool
-
-		wanted string
-	}{
-		"http": {
-			dnsNames: []string{"abc.us-west-1.elb.amazonaws.com"},
-			path:     "svc",
-
-			wanted: "http://abc.us-west-1.elb.amazonaws.com/svc",
-		},
-		"http with / path": {
-			dnsNames: []string{"jobs.test.phonetool.com"},
-			path:     "/",
-
-			wanted: "http://jobs.test.phonetool.com",
-		},
-		"https": {
-			dnsNames: []string{"jobs.test.phonetool.com"},
-			path:     "svc",
-			https:    true,
-
-			wanted: "https://jobs.test.phonetool.com/svc",
-		},
-		"https with / path": {
-			dnsNames: []string{"jobs.test.phonetool.com"},
-			path:     "/",
-			https:    true,
-
-			wanted: "https://jobs.test.phonetool.com",
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			uri := &LBWebServiceURI{
-				DNSNames: tc.dnsNames,
-				Path:     tc.path,
-				HTTPS:    tc.https,
-			}
-
-			require.Equal(t, tc.wanted, uri.String())
-		})
-	}
-}
-
 type lbWebSvcDescriberMocks struct {
 	storeSvc          *mocks.MockDeployedEnvServicesLister
 	ecsStackDescriber *mocks.MockecsStackDescriber
 	envDescriber      *mocks.MockenvDescriber
-}
-
-func TestLBWebServiceDescriber_URI(t *testing.T) {
-	const (
-		testApp          = "phonetool"
-		testEnv          = "test"
-		testSvc          = "jobs"
-		testEnvSubdomain = "test.phonetool.com"
-		testEnvLBDNSName = "abc.us-west-1.elb.amazonaws.com"
-		testSvcPath      = "/"
-	)
-	mockErr := errors.New("some error")
-	testCases := map[string]struct {
-		setupMocks func(mocks lbWebSvcDescriberMocks)
-
-		wantedURI   string
-		wantedError error
-	}{
-		"fail to get parameters of environment stack": {
-			setupMocks: func(m lbWebSvcDescriberMocks) {
-				gomock.InOrder(
-					m.envDescriber.EXPECT().Params().Return(nil, mockErr),
-				)
-			},
-			wantedError: fmt.Errorf("get stack parameters for environment test: some error"),
-		},
-		"fail to get outputs of environment stack": {
-			setupMocks: func(m lbWebSvcDescriberMocks) {
-				gomock.InOrder(
-					m.envDescriber.EXPECT().Params().Return(map[string]string{}, nil),
-					m.envDescriber.EXPECT().Outputs().Return(nil, mockErr),
-				)
-			},
-			wantedError: fmt.Errorf("get stack outputs for environment test: some error"),
-		},
-		"fail to get parameters of service stack": {
-			setupMocks: func(m lbWebSvcDescriberMocks) {
-				gomock.InOrder(
-					m.envDescriber.EXPECT().Params().Return(map[string]string{}, nil),
-					m.envDescriber.EXPECT().Outputs().Return(map[string]string{
-						envOutputPublicLoadBalancerDNSName: testEnvLBDNSName,
-						envOutputSubdomain:                 testEnvSubdomain,
-					}, nil),
-					m.ecsStackDescriber.EXPECT().Params().Return(nil, mockErr),
-				)
-			},
-			wantedError: fmt.Errorf("get stack parameters for service jobs: some error"),
-		},
-		"https web service": {
-			setupMocks: func(m lbWebSvcDescriberMocks) {
-				gomock.InOrder(
-					m.envDescriber.EXPECT().Params().Return(map[string]string{}, nil),
-					m.envDescriber.EXPECT().Outputs().Return(map[string]string{
-						envOutputPublicLoadBalancerDNSName: testEnvLBDNSName,
-						envOutputSubdomain:                 testEnvSubdomain,
-					}, nil),
-					m.ecsStackDescriber.EXPECT().Params().Return(map[string]string{
-						cfnstack.LBWebServiceRulePathParamKey: testSvcPath,
-					}, nil),
-				)
-			},
-
-			wantedURI: "https://jobs.test.phonetool.com",
-		},
-		"http web service": {
-			setupMocks: func(m lbWebSvcDescriberMocks) {
-				gomock.InOrder(
-					m.envDescriber.EXPECT().Params().Return(map[string]string{}, nil),
-					m.envDescriber.EXPECT().Outputs().Return(map[string]string{
-						envOutputPublicLoadBalancerDNSName: testEnvLBDNSName,
-					}, nil),
-					m.ecsStackDescriber.EXPECT().Params().Return(map[string]string{
-						cfnstack.LBWebServiceRulePathParamKey: "*",
-					}, nil),
-				)
-			},
-
-			wantedURI: "http://abc.us-west-1.elb.amazonaws.com/*",
-		},
-		"with alias": {
-			setupMocks: func(m lbWebSvcDescriberMocks) {
-				gomock.InOrder(
-					m.envDescriber.EXPECT().Params().Return(map[string]string{
-						cfnstack.EnvParamAliasesKey: `{"jobs": ["example.com", "v1.example.com"]}`,
-					}, nil),
-					m.envDescriber.EXPECT().Outputs().Return(map[string]string{
-						envOutputPublicLoadBalancerDNSName: testEnvLBDNSName,
-						envOutputSubdomain:                 testEnvSubdomain,
-					}, nil),
-					m.ecsStackDescriber.EXPECT().Params().Return(map[string]string{
-						cfnstack.LBWebServiceRulePathParamKey: testSvcPath,
-					}, nil),
-				)
-			},
-
-			wantedURI: "https://example.com or https://v1.example.com",
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			// GIVEN
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-
-			mockSvcDescriber := mocks.NewMockecsStackDescriber(ctrl)
-			mockEnvDescriber := mocks.NewMockenvDescriber(ctrl)
-			mocks := lbWebSvcDescriberMocks{
-				ecsStackDescriber: mockSvcDescriber,
-				envDescriber:      mockEnvDescriber,
-			}
-
-			tc.setupMocks(mocks)
-
-			d := &LBWebServiceDescriber{
-				ecsServiceDescriber: &ecsServiceDescriber{
-					app: testApp,
-					svc: testSvc,
-					svcStackDescriber: map[string]ecsStackDescriber{
-						"test": mockSvcDescriber,
-					},
-					initDescribers: func(string) error { return nil },
-				},
-
-				envDescriber: map[string]envDescriber{
-					"test": mockEnvDescriber,
-				},
-			}
-
-			// WHEN
-			actual, err := d.URI(testEnv)
-
-			// THEN
-			if tc.wantedError != nil {
-				require.EqualError(t, err, tc.wantedError.Error())
-			} else {
-				require.NoError(t, err)
-				require.Equal(t, tc.wantedURI, actual)
-			}
-		})
-	}
 }
 
 func TestLBWebServiceDescriber_Describe(t *testing.T) {
